trying to fix
The changes you made to your package.json file could indeed be related to the issue.

The caret (^) in front of the version number in your package.json means that when you run npm install, npm will install the newest version that is "compatible" with the specified version according to semantic versioning rules.

However, in practice, this can sometimes lead to issues as "compatible" updates sometimes include changes that break things in subtle ways. This is especially true if the maintainers of the package don't perfectly follow semantic versioning.

In your case, you've changed the react-scripts version from 5.0.1 to ^2.1.3. This is a significant downgrade and likely the cause of the issue. react-scripts includes the Babel configuration and the version you downgraded to may not support the newer JavaScript syntax used by the version of chart.js you're using.

The change in @babel/types version from ^7.22.3 to ^7.22.5 is less significant since it's only a patch version increase, which should not introduce breaking changes.

I would suggest trying to revert the react-scripts version back to 5.0.1, run npm install again, and then try to build your project:

Open your package.json file.
Change the line with "react-scripts": "^2.1.3" back to "react-scripts": "5.0.1".
Save the file.
Run npm install in the terminal in your project directory.
Try building your project again.
If this doesn't work, we'll try downgrading chart.js to an earlier version. For this, I'll need the original chart.js version number before it got updated.