Fix og:image URL for Grower Profile Metadata
This commit addresses the issue where the og:image URL in the metadata for grower profiles was not being processed correctly due to the use of %PUBLIC_URL%. The problem was identified when sharing grower profile pages on social media platforms, which failed to display the intended image.
Changes made:
Removed the %PUBLIC_URL% placeholder from the og:image URL construction within the generateMetadata function in app/grower/[growerName]/page.js.
Introduced a new environment variable NEXT_PUBLIC_BASE_URL to store the base URL of the application (https://www.pumpkinpal.app). This variable is now used to construct the full URL for og:image, ensuring it is dynamically set based on the deployment environment.
Updated the og:image URL to use NEXT_PUBLIC_BASE_URL, concatenating it with the path to the image, ensuring the URL is correctly formed and accessible by social media platforms when fetching the image.
This fix ensures that the og:image is correctly displayed when grower profile pages are shared, enhancing the shareability and visual representation of the content on social platforms.