modify the calculateGlobalRankings
This updated function will create two new fields in each pumpkin document:
lifetimeGlobalRanks: An object with keys for each category and their corresponding lifetime ranks.
yearGlobalRanks: An object with keys for each category and their corresponding yearly ranks.
This structure is more in line with our original plan and will make it easier to access and update rankings for different categories.